Remove more references to cost
